test(engine): guard kwargs and split coverage
